- Updated parts of trunk to be compatible with Visual Studio (ie, get rid of float/double warnings - Added massive amount of files for ProductionCodeTester, it all compiles now
git-svn-id: https://kwarqs-dev.bbn.com/svn/kwarqs/code@217 eb211378-2a44-443b-ae4a-d45323d45073

+++ ProductionCode/trunk/Framework/KwarqsGamePiece.cpp
@@ -24,10 +24,10 @@ void KwarqsGamePiece::PerformMovement()
 	else
 		m_low_arm_motor.Set(0);
 	
-	double val = ((m_stick->GetZ()*-1) + 1.0)/ 2;
+	float val = ((m_stick->GetZ()*-1.0F) + 1.0F)/ 2.0F;
 	
 	// scale this value so it can only go the correct direction
-	m_upper_arm_motor.Set(val*-1);
+	m_upper_arm_motor.Set(val*-1.0F);
 	
 	DriverStationLCD::GetInstance()->Printf(DriverStationLCD::kUser_Line6, 1, "%f", val);
 }
